refactor: move scrubbingLogger to the scrubber pkg
We need it exported so we can use it in the new implementation.

Part of https://github.com/ooni/probe/issues/1687
